refactor: extract populateCategoryFilter into shared module (#78)
Extract the duplicated DOM manipulation (clearing and repopulating a
<select> filter) into modules/ui-helpers.js as setFilterOptions().

- options/options.js: populateCategoryFilter() is now a one-liner that
  delegates DOM work to setFilterOptions()
- popup/popup.js: identity-aware variant likewise delegates its DOM
  work to setFilterOptions(); also removes the now-unused getCategories
  import that was left over from an earlier version of the function
